refactor to use unsigned integers for sizes
resolves -Wstringop-overflow

size_t

+++ lcm/lcm_coretypes.h
@@ -119,8 +119,8 @@ static inline int __int8_t_encode_array(void *_buf, int offset, int maxlen, cons
     return elements;
 }
 
-static inline int __int8_t_decode_array(const void *_buf, int offset, int maxlen, int8_t *p,
-                                        int elements)
+static inline int __int8_t_decode_array(const void *_buf, size_t offset, size_t maxlen, int8_t *p,
+                                        size_t elements)
 {
     if (maxlen < elements)
         return -1;
